+++ lib/src/flutter_branch_sdk_platform_interface.dart
@@ -30,12 +30,10 @@ abstract class FlutterBranchSdkPlatform extends PlatformInterface {
   ///Initialize Branch SDK
   /// [useTestKey] - Sets `true` to use the test `key_test_...
   /// [enableLogging] - Sets `true` turn on debug logging
-  /// [enableFacebookLinkCheck] - Sets `true` to enable Facebook app link check operation during Branch initialisation
   /// [disableTracking] - Sets `true` to disable tracking in Branch SDK for GDPR compliant on start. After having consent, sets `false`
   Future<void> init(
       {bool useTestKey = false,
       bool enableLogging = false,
-      bool enableFacebookLinkCheck = false,
       bool disableTracking = false}) async {
     throw UnimplementedError('init has not been implemented');
   }
@@ -138,14 +136,6 @@ abstract class FlutterBranchSdkPlatform extends PlatformInterface {
       {required BranchUniversalObject buo,
       BranchLinkProperties? linkProperties}) async {
     throw UnimplementedError('removeFromSearch has not been implemented');
-  }
-
-  ///Set time window for SKAdNetwork callouts in Hours (Only iOS)
-  ///By default, Branch limits calls to SKAdNetwork to within 72 hours after first install.
-  @Deprecated('This is no longer supported for iOS 16.1+')
-  void setIOSSKAdNetworkMaxTime(int hours) {
-    throw UnimplementedError(
-        'setIOSSKAdNetworkMaxTime has not been implemented');
   }
 
   ///Indicates whether or not this user has a custom identity specified for them. Note that this is independent of installs.