+++ lib/src/flutter_branch_sdk_web.dart
@@ -34,13 +34,11 @@ class FlutterBranchSdkWeb extends FlutterBranchSdkPlatform {
   ///Initialize Branch SDK
   /// [useTestKey] - Sets `true` to use the test `key_test_...
   /// [enableLogging] - Sets `true` turn on debug logging
-  /// [enableFacebookLinkCheck] - Sets `true` to enable Facebook app link check operation during Branch initialisation
   /// [disableTracking] - Sets `true` to disable tracking in Branch SDK for GDPR compliant on start. After having consent, sets `false`
   @override
   Future<void> init(
       {bool useTestKey = false,
       bool enableLogging = false,
-      bool enableFacebookLinkCheck = false,
       bool disableTracking = false}) async {
     debugPrint('');
   }
@@ -292,15 +290,6 @@ class FlutterBranchSdkWeb extends FlutterBranchSdkPlatform {
       {required BranchUniversalObject buo,
       BranchLinkProperties? linkProperties}) async {
     throw UnsupportedError('removeFromSearch() Not supported by Branch JS SDK');
-  }
-
-  ///Set time window for SKAdNetwork callouts in Hours (Only iOS)
-  ///By default, Branch limits calls to SKAdNetwork to within 72 hours after first install.
-  @Deprecated('This is no longer supported for iOS 16.1+')
-  @override
-  void setIOSSKAdNetworkMaxTime(int hours) {
-    throw UnsupportedError(
-        'setIOSSKAdNetworkMaxTime() Not available in Branch JS SDK');
   }
 
   ///Indicates whether or not this user has a custom identity specified for them. Note that this is independent of installs.